refactor(tests): split runtime UI state coverage
